# order-service/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Order Service API
  description: API сервиса заказов
  version: 0.1.0
servers:
  - url: http://localhost:8084/orders
    description: Dev server

paths:
  /cart/me:
    get:
      summary: Получить мою корзину
      description: Возвращает список L3-доменов в корзине текущего пользователя с расчетом цен. Требуется JWT.
      operationId: getCartMe
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о корзине
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /cart/{l3Domain}:
    post:
      summary: Добавить L3-домен в корзину
      description: Добавляет домен третьего уровня в корзину текущего пользователя. Требуется JWT.
      operationId: addToCart
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: l3Domain
          in: path
          required: true
          description: Имя домена третьего уровня
          schema:
            type: string
            pattern: ^\s*([a-zA-Z0-9-]+\.){2}[a-zA-Z]{2,}\s*$
            example: iva.example.com
      responses:
        '201':
          description: Добавлено в корзину
        '400':
          description: Некорректные входные данные. Возвращает список атрибутов с ошибками.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /cart/checkout:
    post:
      summary: Оформить заказ
      description: Собирает все домены из корзины текущего пользователя, отправляет их в domain-service для регистрации с указанным периодом, после успешного ответа очищает корзину. Требуется JWT.
      operationId: checkout
      tags:
        - Cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Заказ успешно оформлен. Возвращает список зарегистрированных доменов.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Зарегистрированный домен третьего уровня
                  example: "iva.example.com"
        '400':
          description: Корзина пуста или произошла ошибка.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /domains/renew:
    post:
      summary: Продлить домены
      description: Продлевает срок действия указанных доменов третьего уровня на выбранный период. Требуется JWT.
      operationId: renewDomains
      tags:
        - Domain
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewDomainsRequest'
      responses:
        '200':
          description: Домены успешно продлены.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Продлённый домен третьего уровня
                  example: "iva.example.com"
        '400':
          description: Некорректные входные данные.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DomainPeriod:
      type: string
      enum:
        - MONTH
        - YEAR
      description: Период действия домена
    CheckoutRequest:
      type: object
      required:
        - period
      properties:
        period:
          $ref: '#/components/schemas/DomainPeriod'
    RenewDomainsRequest:
      type: object
      required:
        - l3Domains
        - period
      properties:
        l3Domains:
          type: array
          items:
            type: string
            pattern: ^\s*([a-zA-Z0-9-]+\.){2}[a-zA-Z]{2,}\s*$
            example: "iva.example.com"
          minItems: 1
          description: Список доменов третьего уровня для продления
        period:
          $ref: '#/components/schemas/DomainPeriod'
    CartResponse:
      type: object
      required:
        - totalMonthlyPrice
        - totalYearlyPrice
        - l3Domains
      properties:
        totalMonthlyPrice:
          type: integer
          format: int32
          description: Суммарная цена всех доменов за месяц
        totalYearlyPrice:
          type: integer
          format: int32
          description: Суммарная цена всех доменов за год (с учетом скидки)
        l3Domains:
          type: array
          items:
            type: string
          description: Список L3-доменов в корзине